+++ extension/src/webview/views/GitCredentials/index.ts
@@ -1,1 +1,0 @@
-export { GitCredentialsView } from './GitCredentialsView';