+++ extension/src/webview/views/SubmissionSetup/index.ts
@@ -1,0 +1,1 @@
+export { SubmissionSetupView } from './SubmissionSetupView';